requirements.txt
pyyaml
numpy>=2.2.3
open3d